Inventory templates and more stringent views

+++ inventory/views.py
@@ -1,14 +1,21 @@
-from django.shortcuts import render, HttpResponseRedirect, get_object_or_404, HttpResponse
+from django.shortcuts import render, HttpResponseRedirect, get_object_or_404
 from django.core.urlresolvers import reverse
 from django.contrib import messages
 from .models import Asset
 from .forms import AssetForm
 from courselib.auth import requires_role
 from log.models import LogEntry
-from coredata.models import Unit
+from coredata.models import Unit, Role
+from courselib.auth import ForbiddenResponse
 import unicodecsv as csv
 from datetime import datetime
 
+
+def _has_unit_role(user, asset):
+    """
+    A quick method to check that the person has the Inventory Admin role for the given asset's unit.
+    """
+    return Role.objects.filter(person__userid=user.username, role='INV', unit=asset.unit).count() > 0
 
 @requires_role('INV')
 def index(request):
@@ -40,6 +47,8 @@ def new_asset(request):
 @requires_role('INV')
 def edit_asset(request, asset_slug):
     asset = get_object_or_404(Asset, slug=asset_slug)
+    if not _has_unit_role(request.user, asset):
+        return ForbiddenResponse(request)
     if request.method == 'POST':
         form = AssetForm(request, request.POST, instance=asset)
         if form.is_valid():
@@ -54,18 +63,22 @@ def edit_asset(request, asset_slug):
             return HttpResponseRedirect(reverse(index))
     else:
         form = AssetForm(request, instance=asset)
-    return render(request, 'inventory/edit_asset.html', {'form': form})
+    return render(request, 'inventory/edit_asset.html', {'form': form, 'asset_slug': asset_slug})
 
 
 @requires_role('INV')
 def view_asset(request, asset_slug):
     asset = get_object_or_404(Asset, slug=asset_slug)
+    if not _has_unit_role(request.user, asset):
+        return ForbiddenResponse(request)
     return render(request, 'inventory/view_asset.html', {'asset': asset})
 
 
 @requires_role('INV')
 def delete_asset(request, asset_id):
     asset = get_object_or_404(Asset, pk=asset_id)
+    if not _has_unit_role(request.user, asset):
+        return ForbiddenResponse(request)
     if request.method == 'POST':
         asset.delete()
         messages.success(request, 'Hid asset %s' % asset)
